# Refactor and speed up Mailer specs