--- a/app/views/candidate_mailer/ucas_match_resolved_on_ucas_email.text.erb
+++ b/app/views/candidate_mailer/ucas_match_resolved_on_ucas_email.text.erb
@@ -1,8 +1,8 @@
 Dear <%= @application_form.full_name %>
 
-We’ve written to you a couple of times as we noticed you applied to study  <%= @course_name_and_code %> at <%= @provider_name %> on both Apply for teacher training and UCAS Teacher Training Apply – which isn’t allowed.
+We’ve written to you a couple of times as we noticed you applied to study <%= @course_name_and_code %> at <%= @provider_name %> on both Apply for teacher training and UCAS Teacher Training Apply – which isn’t allowed.
 
-As you didn’t withdraw the choice, as requested, we have removed it from your UCAS application meaning you should use Apply for teacher training to track the progress of your application for [course name and code].
+As you didn’t withdraw the choice, as requested, we have removed it from your UCAS application meaning you should use Apply for teacher training to track the progress of your application for <%= @course_name_and_code %>.
 
 Any other courses you applied to have not been affected.