--- a/spec/mailers/candidate_mailer_ucas_match_reminder_email_spec.rb
+++ b/spec/mailers/candidate_mailer_ucas_match_reminder_email_spec.rb
@@ -1,8 +1,6 @@
 require 'rails_helper'
 
 RSpec.describe CandidateMailer, type: :mailer do
-  include TestHelpers::MailerSetupHelper
-
   around do |example|
     Timecop.freeze(Date.new(2020, 11, 23)) do
       example.run
@@ -11,59 +9,38 @@ RSpec.describe CandidateMailer, type: :mailer do
 
   subject(:mailer) { described_class }
 
-  let(:application_choice) { create(:application_choice) }
-  let(:application_form) { create(:completed_application_form, application_choices: [application_choice]) }
-  let(:ucas_match) { create(:ucas_match, :need_to_send_reminder_emails, application_form: application_form) }
+  let(:ucas_match) { build_stubbed(:ucas_match, candidate_last_contacted_at: Time.zone.local(2020, 11, 16)) }
+  let(:application_form) { build_stubbed(:application_form, first_name: 'Jane', application_choices: [application_choice]) }
+  let(:provider) { build_stubbed(:provider, name: 'City University') }
+  let(:course_option) { build_stubbed(:course_option, course: course) }
+  let(:course) { build_stubbed(:course, name: 'Physics', code: '3PH5', provider: provider) }
+  let(:application_choice) { build_stubbed(:application_choice, course_option: course_option) }
 
   describe '.ucas_match_reminder_email_duplicate_applications' do
-    let(:email) { mailer.ucas_match_reminder_email_duplicate_applications(application_choice, ucas_match) }
+    let(:email) { mailer.ucas_match_reminder_email_duplicate_applications(application_form.application_choices.first, ucas_match) }
 
-    it 'sends an email with the correct subject' do
-      expect(email.subject).to include(I18n.t!('candidate_mailer.ucas_match_reminder_email.duplicate_applications.subject'))
-    end
-
-    it 'sends an email with the correct heading' do
-      expect(email.body.encoded).to include("Dear #{application_form.full_name}")
-    end
-
-    it 'sends an email containing the course name and code in the body' do
-      course_name_and_code = application_choice.course.name_and_code
-
-      expect(email.body).to include(course_name_and_code)
-    end
-
-    it 'sends an email containing the provider name in the body' do
-      provider_name = application_choice.course.provider.name
-
-      expect(email.body).to include(provider_name)
-    end
-
-    it 'sends an email containing the date the initial email was sent' do
-      expect(email.body).to include('16 November 2020')
-    end
-
-    it 'sends an email containing the date that the application needs to be withdrawn by' do
-      expect(email.body).to include('30 November 2020')
-    end
+    it_behaves_like(
+      'a mail with subject and content',
+      I18n.t!('candidate_mailer.ucas_match_reminder_email.duplicate_applications.subject'),
+      'heading' => 'Dear Jane',
+      'course name and code' => 'Physics (3PH5)',
+      'provider' => 'City University',
+      'initial email date' => '16 November 2020',
+      'withdrawn by date' => '30 November 2020',
+    )
   end
 
   describe '.ucas_match_reminder_email_multiple_acceptances' do
-    let(:email) { mailer.ucas_match_reminder_email_multiple_acceptances(ucas_match) }
+    let(:email) { mailer.ucas_match_reminder_email_multiple_acceptances(candidate.ucas_match) }
+    let(:ucas_match) { build_stubbed(:ucas_match, candidate_last_contacted_at: Time.zone.local(2020, 11, 16)) }
+    let(:candidate) { build_stubbed(:candidate, ucas_match: ucas_match, application_forms: [application_form]) }
 
-    it 'sends an email with the correct subject' do
-      expect(email.subject).to include(I18n.t!('candidate_mailer.ucas_match_reminder_email.multiple_acceptances.subject'))
-    end
-
-    it 'sends an email with the correct heading' do
-      expect(email.body.encoded).to include("Dear #{application_form.full_name}")
-    end
-
-    it 'sends an email containing the date the initial email was sent' do
-      expect(email.body).to include('16 November 2020')
-    end
-
-    it 'sends an email containing the date that the application needs to be withdrawn by' do
-      expect(email.body).to include('30 November 2020')
-    end
+    it_behaves_like(
+      'a mail with subject and content',
+      I18n.t!('candidate_mailer.ucas_match_reminder_email.multiple_acceptances.subject'),
+      'heading' => 'Dear Jane',
+      'initial email date' => '16 November 2020',
+      'withdrawn by date' => '30 November 2020',
+    )
   end
 end