… spec/mailers/candidate_mailer_offers_and_rejections_spec.rb
… spec/mailers/candidate_mailer_spec.rb
… spec/mailers/provider_mailer_spec.rb
… spec/mailers/referee_mailer_spec.rb