## Context

- Extract reusable shared_example and group checks (this significantly reduces the spec running time as it **reduces the mailer tests from 221 to 55** just by grouping the checks) 
- Refactor specs to enable usage of the shared example (and make sure we only use stubbed data)
- Disable rubocop RSpec/LetSetup to make it easier to setup data using let! (This enables using the parameter directly instead of having to for example, go through the application form to access the references `application_form.application_references.first` to ensure that the application_form object is created)
- includes a fix for the resolved_ucas_email content (remove space, display course name and code 

## Before

> Finished in 19.63 seconds (files took 11.98 seconds to load)
> 221 examples, 0 failures

## After

> Finished in 4.29 seconds (files took 11.97 seconds to load)
> 55 examples, 0 failures